WIP split cache classes onto a separate module.
